refactor: move search implementation out of strategies store
