+++ src/lib/features/feature-search/fake-feature-search-store.ts
@@ -1,0 +1,15 @@
+import { IFeatureOverview } from 'lib/types';
+import {
+    IFeatureSearchParams,
+    IQueryParam,
+} from '../feature-toggle/types/feature-toggle-strategies-store-type';
+import { IFeatureSearchStore } from './feature-search-store-type';
+
+export default class FakeFeatureSearchStore implements IFeatureSearchStore {
+    searchFeatures(
+        params: IFeatureSearchParams,
+        queryParams: IQueryParam[],
+    ): Promise<{ features: IFeatureOverview[]; total: number }> {
+        throw new Error('Method not implemented.');
+    }
+}
