+++ src/lib/features/feature-search/feature-search-store-type.ts
@@ -1,0 +1,15 @@
+import {
+    IFeatureSearchParams,
+    IQueryParam,
+} from '../feature-toggle/types/feature-toggle-strategies-store-type';
+import { IFeatureOverview } from '../../types';
+
+export interface IFeatureSearchStore {
+    searchFeatures(
+        params: IFeatureSearchParams,
+        queryParams: IQueryParam[],
+    ): Promise<{
+        features: IFeatureOverview[];
+        total: number;
+    }>;
+}
